改结构

+++ src/main/java/com/example/game_shop/service/GameService.java
@@ -1,13 +1,18 @@
 package com.example.game_shop.service;
 
+import com.example.game_shop.Result.Result;
 import com.example.game_shop.mapper.GameMapper;
 import com.example.game_shop.mapper.GamePictureMapper;
+import com.example.game_shop.mapper.UserMapper;
 import com.example.game_shop.pojo.BasicGameInfo;
 import com.example.game_shop.pojo.Game;
+import com.example.game_shop.utils.ResultUtil;
+import com.example.game_shop.utils.TokenUtil;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 /**
@@ -17,10 +22,13 @@ import java.util.List;
 @Service
 public class GameService {
     @Resource
-    GameMapper gameMapper;
+    private GameMapper gameMapper;
 
     @Resource
-    GamePictureMapper pictureMapper;
+    private TokenUtil tokenUtil;
+
+    @Resource
+    private GamePictureMapper pictureMapper;
 
     /**
      * 游戏是否存在
@@ -31,24 +39,41 @@ public class GameService {
         return gameMapper.getGameById(id) != null;
     }
 
-    public Game getGameById(long id) {
-        return gameMapper.getGameById(id);
+    /**
+     * 获取游戏的详细信息
+     */
+    public Result<Game> getGameById(long id) {
+        return ResultUtil.success(gameMapper.getGameById(id));
     }
 
-    public List<BasicGameInfo> queryGame(String name, String platform) {
-        return gameMapper.queryGame(name, platform);
+    /**
+     * 查询游戏
+     *
+     * @param name     游戏名
+     * @param platform 游戏平台
+     * @param page     页号
+     */
+    public Result<List<BasicGameInfo>> queryGame(String name, String platform, int page) {
+        List<BasicGameInfo> gameList = gameMapper.queryGame(name, platform);
+        // 步长20
+        int step = 20;
+        return ResultUtil.success(gameList.subList((page - 1) * step, (page * step)));
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public void updateGameInfo(Game game) {
-        int updated = gameMapper.updateGameInfo(game);
-        int deleted = pictureMapper.deletePicture(game.getId());
-        int inserted = pictureMapper.addPicture(game.getId(), game.getImages());
+    public Result<String> updateGameInfo(Game game, HttpServletRequest request) {
+        if ("管理员".equals(tokenUtil.getIdent(request))) {
+            int updated = gameMapper.updateGameInfo(game);
+            int deleted = pictureMapper.deletePicture(game.getId());
+            int inserted = pictureMapper.addPicture(game.getId(), game.getImages());
 
-        System.out.println("update game: " + updated);
-        System.out.println("delete picture: " + deleted);
-        System.out.println("insert picture: " + inserted);
+            System.out.println("update game: " + updated);
+            System.out.println("delete picture: " + deleted);
+            System.out.println("insert picture: " + inserted);
+
+            return ResultUtil.success("修改成功", null);
+        } else {
+            return ResultUtil.fail("权限不足");
+        }
     }
-
-
 }
