+++ src/main/java/com/example/game_shop/service/RecyclePackService.java
@@ -1,12 +1,16 @@
 package com.example.game_shop.service;
 
+import com.example.game_shop.Result.Result;
 import com.example.game_shop.mapper.GameMapper;
 import com.example.game_shop.mapper.RecyclePackMapper;
 import com.example.game_shop.pojo.RecyclePackGame;
+import com.example.game_shop.utils.ResultUtil;
+import com.example.game_shop.utils.TokenUtil;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.Map;
 
@@ -17,34 +21,58 @@ import java.util.Map;
 @Service
 public class RecyclePackService {
     @Resource
-    RecyclePackMapper packMapper;
+    private RecyclePackMapper packMapper;
 
     @Resource
-    GameMapper gameMapper;
+    private GameMapper gameMapper;
+
+    @Resource
+    private TokenUtil tokenUtil;
 
     public boolean hasGame(String account, long id, String type) {
         return packMapper.hasGame(account, id, type);
     }
 
-    public List<RecyclePackGame> getGames(String account, String type) {
-        return packMapper.getGames(account, type);
+
+    public Result<List<RecyclePackGame>> getGames(String account, String type, HttpServletRequest request) {
+        if (tokenUtil.check(account, request)) {
+            return ResultUtil.success(packMapper.getGames(account, type));
+        } else {
+            return ResultUtil.fail("用户不一致");
+        }
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public void deleteGames(String account, String type, List<Long> idList) {
-        int deleteNum = packMapper.deleteGames(account, type, idList);
-        System.out.println("delete num:" + deleteNum);
+    public Result<String> deleteGames(String account, String type, List<Long> idList, HttpServletRequest request) {
+        if (tokenUtil.check(account, request)) {
+            int deleted = packMapper.deleteGames(account, type, idList);
+            System.out.println("delete num:" + deleted);
+            return ResultUtil.success("删除成功", null);
+        } else {
+            return ResultUtil.fail("用户不一致");
+        }
+    }
+
+
+    @Transactional(rollbackFor = Exception.class)
+    public Result<String> updateNum(String account, String type, List<Map<String, Object>> numList, HttpServletRequest request) {
+        if (tokenUtil.check(account, request)) {
+            int updated = packMapper.updateNum(account, type, numList);
+            System.out.println("update recycle_pack_game: " + updated);
+            return ResultUtil.success("删除成功", null);
+        } else {
+            return ResultUtil.fail("用户不一致");
+        }
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public void updateNum(String account, String type, List<Map<String, Object>> numList) {
-        int updated = packMapper.updateNum(account, type, numList);
-        System.out.println("update recycle_pack_game: " + updated);
-    }
-
-    @Transactional(rollbackFor = Exception.class)
-    public void addGame(String account, long gid, String type) {
-        int gameAdded = packMapper.addGame(account, gid, gameMapper.getPrice(gid), type);
-        System.out.println("add num:" + gameAdded);
+    public Result<String> addGame(String account, long gid, String type, HttpServletRequest request) {
+        if (tokenUtil.check(account, request)) {
+            int added = packMapper.addGame(account, gid, gameMapper.getPrice(gid), type);
+            System.out.println("add num:" + added);
+            return ResultUtil.success("删除成功", String.valueOf(added));
+        } else {
+            return ResultUtil.fail("用户不一致");
+        }
     }
 }
